# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AutoAssist IA API
  version: 1.0.0
  description: API do AutoAssist - Ecossistema automotivo com IA. Consulte os endpoints
    para chat, manutenção preditiva, FIPE, pagamentos e mais.
servers:
- url: http://localhost:5001
  description: Desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
- bearerAuth: []
paths:
...
  /api/veiculos:
    get:
      summary: Listar veiculos do usuario
      tags:
      - Veiculos
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Lista de veiculos
    post:
      summary: Adicionar veiculo
      tags:
      - Veiculos
      security:
      - bearerAuth: []
      responses:
        '201':
          description: Veiculo adicionado
  /api/veiculos/{id}/foto:
    post:
      summary: Upload da foto do veículo (JSON base64 legado ou multipart)
      tags:
      - Veiculos
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Foto salva (retorna foto_url + foto_base64 legado quando aplicável)
    get:
      summary: Serve a foto do veículo (redirect S3 ou arquivo local)
      tags:
      - Veiculos
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Imagem
        '302':
          description: Redirect para storage externo
...